Refactored the draw method in the DisplayEntityHealthComponent class (formally WallStatsDisplay) to apply a reduction factor to the current scale instead of relying on magic numbers which could become outdated when a new map is implemented.
