+++ source/core/src/main/com/csse3200/game/components/DisplayEntityHealthComponent.java
@@ -1,24 +1,24 @@
-package com.csse3200.game.components.buildables;
+package com.csse3200.game.components;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import com.badlogic.gdx.math.Matrix3;
 import com.badlogic.gdx.math.Matrix4;
+import com.badlogic.gdx.math.Quaternion;
+import com.badlogic.gdx.math.Vector3;
 import com.badlogic.gdx.scenes.scene2d.ui.Image;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.csse3200.game.components.CombatStatsComponent;
-import com.csse3200.game.entities.Entity;
 import com.csse3200.game.rendering.RenderComponent;
 import com.csse3200.game.services.ServiceLocator;
 import com.csse3200.game.ui.UIComponent;
 
-import java.awt.*;
-
 /**
  * A ui component for displaying a walls health on top of its location on the map.
  */
-public class WallStatsDisplay extends UIComponent {
-    private static final float SCALE_REDUCTION = 1000;
+public class DisplayEntityHealthComponent extends UIComponent {
+    private static final float SCALE_REDUCTION = 70;
     Table table;
     private Image heartImage;
     private Label healthLabel;
@@ -71,12 +71,23 @@ public class WallStatsDisplay extends UIComponent {
     @Override
     public void draw(SpriteBatch batch)  {
         Matrix4 originalMatrix = batch.getProjectionMatrix().cpy(); // cpy() needed to properly set afterwards because calling set() seems to modify kept matrix, not replaces it
-        batch.setProjectionMatrix(originalMatrix.cpy().scale(getWidthUnitsInPixel(), getHeightUnitsInPixel(), 1));
+
+        var newScale = new Vector3();
+        newScale = originalMatrix.getScale(newScale);
+        newScale.scl(1/SCALE_REDUCTION);
+
+        var originalPosition = originalMatrix.getTranslation(new Vector3());
+        var originalRotation = originalMatrix.getRotation(new Quaternion());
+
+        var newMatrix = new Matrix4(originalPosition, originalRotation, newScale);
+        batch.setProjectionMatrix(newMatrix);
         var old_x = table.getX();
         var old_y = table.getY();
 
+        var position = new Vector3(old_x, old_y, 0);
+        position.scl(SCALE_REDUCTION);
         // scale the positions to match the new projection matrix
-        table.setPosition(old_x * SCALE_REDUCTION / 20f, old_y * SCALE_REDUCTION / 20f);
+        table.setPosition(position.x, position.y);
         table.draw(batch, 1.0f);
         table.setPosition(old_x, old_y);
         batch.setProjectionMatrix(originalMatrix); //revert projection
